:recycle: #51 - major module refactoring
