+++ src/modules/major/entities/main-skill.entity.ts
