+++ src/modules/major/services/major.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { MainSkill } from '@prisma/client';
 import { PrismaService } from '@src/modules/core/database/prisma/prisma.service';
+import { MainSkillEntity } from '@src/modules/major/entities/main-skill.entity';
 import { MajorEntity } from '@src/modules/major/entities/major.entity';
 import { MajorRelationFieldRequestQueryDto } from '../dto/major-relation-field-request-query.dto';
 
@@ -8,6 +8,10 @@ import { MajorRelationFieldRequestQueryDto } from '../dto/major-relation-field-r
 export class MajorService {
   constructor(private readonly prismaService: PrismaService) {}
 
+  /**
+   * major 전체조회
+   * 클라언트에게 받은 mainSKills: boolean 를 가지고 mainSkills 도 함께 가져온다.
+   */
   findMajors(query: MajorRelationFieldRequestQueryDto): Promise<MajorEntity[]> {
     const { mainSkills } = query;
 
@@ -18,6 +22,10 @@ export class MajorService {
     });
   }
 
+  /**
+   * major 단일 조회
+   * 클라언트에게 받은 mainSKills: boolean 를 가지고 mainSkills 도 함께 가져온다.
+   */
   findMajor(
     majorId: number,
     query: MajorRelationFieldRequestQueryDto,
@@ -34,7 +42,11 @@ export class MajorService {
     });
   }
 
-  findMainSkills(majorId: number): Promise<MainSkill[]> {
+  /**
+   * mainSkill 전체 조회
+   * majorId 를 기준으로 mainSkill 을 가져온다.
+   */
+  findMainSkills(majorId: number): Promise<MainSkillEntity[]> {
     return this.prismaService.mainSkill.findMany({
       where: {
         majorId,
@@ -42,7 +54,14 @@ export class MajorService {
     });
   }
 
-  findMainSkill(majorId: number, mainSkillId: number): Promise<MainSkill> {
+  /**
+   * mainSkill 단일 조회
+   * majorId 를 기준으로 mainSkill 을 가져온다.
+   */
+  findMainSkill(
+    majorId: number,
+    mainSkillId: number,
+  ): Promise<MainSkillEntity> {
     return this.prismaService.mainSkill.findUnique({
       where: {
         majorId,
